model/requirements.txt
# Core ML dependencies
numpy>=1.24.0,<2.0.0

torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
scikit-learn>=1.3.0,<2.0.0
pyyaml>=6.0,<7.0

# Optional (for data analysis and visualization)
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
